refactoring teleop
